requirements.txt
streamlit
nltk
scikit-learn
pandas
PyPDF2
python-dotenv
google-genai
